requirements.txt
pandas==2.2.2
openai==1.55.3
httpx==0.27.2
langchain==0.1.16
langchain-community==0.0.36
langchain-openai==0.1.7
python-dotenv==1.0.1
tqdm==4.66.4
typing-extensions>=4.11.0,<5.0.0
